react 16 support added
Forgot I switched to a private repo, so updating directly from github.

Small change to support react 16, which deprecated the onTouchTap property.
Additionally the boolean value passed to autoCapitalize is invalid, needs to be a string, which works fine on react v 15 too. 
Added a version check to determine which Button attribute-set to utilize. 
React 16 doesn't support the inject tap event plugin. I switched to fastclick and haven't had any problems. 
Readme.md could be updated, but is anyone even using this?